Moved _close_fp into ImageFile
