+++ src/PIL/Image.py
@@ -603,16 +603,10 @@ class Image:
     def __enter__(self):
         return self
 
-    def _close_fp(self):
-        if getattr(self, "_fp", False):
-            if self._fp != self.fp:
-                self._fp.close()
-            self._fp = DeferredError(ValueError("Operation on closed image"))
-        if self.fp:
-            self.fp.close()
-
     def __exit__(self, *args):
-        if hasattr(self, "fp"):
+        from . import ImageFile
+
+        if isinstance(self, ImageFile.ImageFile):
             if getattr(self, "_exclusive_fp", False):
                 self._close_fp()
             self.fp = None
